// source/packages/@aws-accelerator/modules/models/enums.ts
export enum AcceleratorModules {
  /**
   * AWS Control Tower Landing Zone module setup module
   *
   * @description
   * For more information on WS Service control policies, please refer the [document](https://docs.aws.amazon.com/controltower/latest/userguide/what-is-control-tower.html)
   */
  SETUP_CONTROL_TOWER_LANDING_ZONE = 'control-tower-landing-zone',
  /**
   * AWS Organizations Organizational Unit (OU) create module
   */
  CREATE_ORGANIZATIONAL_UNIT = 'create-organizational-unit',
  /**
   * Register AWS Organizations Organizational Unit (OU) with AWS Control Tower module
   */
  REGISTER_ORGANIZATIONAL_UNIT = 'register-organizational-unit',
  /**
   * Invite AWS Accounts to AWS Organizations module
   */
  INVITE_ACCOUNTS_TO_ORGANIZATIONS = 'invite-accounts-to-organizations',
  /**
   * Move AWS Accounts to destination AWS Organizations Organizational Unit (OU) module
   */
  MOVE_ACCOUNTS = 'move-accounts',
  /**
   * Retrieves cross account CloudFormation Templates
   */
  GET_CLOUDFORMATION_TEMPLATES = 'get-cloudformation-templates',

  /**
   * Set stack policy for accounts.
   */
  CREATE_STACK_POLICY = 'create-stack-policy',

  /**
   * Configure IAM Root User Management
   */

  ROOT_USER_MANAGEMENT = 'root-user-management',

  /**
   * SSM Block Public Document Sharing module
   *
   * @description
   * For more information on SSM Block Public Document Sharing, please refer the [document](https://docs.aws.amazon.com/systems-manager/latest/userguide/document-public-access-block.html)
   */
  SSM_BLOCK_PUBLIC_DOCUMENT_SHARING = 'ssm-block-public-document-sharing',
  /**
   * Sets an account alias
   */
  MANAGE_ACCOUNTS_ALIAS = 'manage-accounts-alias',

  /**
   * Manages AWS Organizations tags on accounts
   */
  MANAGE_ACCOUNTS_TAGS = 'manage-accounts-tags',

  /**
   * Accelerator prerequisites module
   */
  ACCELERATOR_PREREQUISITES = 'accelerator-prerequisites',

  /**
   * Pipeline prerequisites module
   */
  PIPELINE_PREREQUISITES = 'pipeline-prerequisites',

  /**
   * Delete default VPCs in accounts
   */
  DELETE_DEFAULT_VPC = 'delete-default-vpc',

  /**
   * Manage Security Hub Automation rules module
   */
  MANAGE_AUTOMATION_RULES = 'manage-automation-rules',

  /**
   * Manage Control Tower accounts enrollment
   */
  ENROLL_ACCOUNTS = 'enroll-accounts',

  /**
   * An Example module which is executed in `PREPARE` stage
   */
  EXAMPLE_MODULE = 'example-module',
}

/**
 * Module execution phase
 *
 * @description
 * This is the list of module execution phases
 */
export enum ModuleExecutionPhase {
  /**
   * At the synth phase during bootstrap stage of accelerator pipeline
   */
  SYNTH = 'synth',
  /**
   * At the deploy phase during of accelerator pipeline stage
   */
  DEPLOY = 'deploy',
}